Catch key presses
It is left as a remainder so I don't have to go and dig it up
in the GTK docs when I get around implementing stuff.

It will need a dispatcher for every feature... might be even
hookable, who knows.

Stuff needs to be thread safe, it is easy because most of the
stuff will just block.

Signed-off-by: Fernando J. Pereda <ferdy@gentoo.org>